Add a way to refer back to a common ancestor
This allows querying for a potentially far-off ancestor in the bevy
hierarchy with the `Ancestor` component.

It's a very simple system that recursively traverse the children of an
entity marked with a `MarkDescendants` component, and adds a `Ancestor`
component to entities that comply to a rule provided in the
`MarkDescendants` component.

drawbacks:
- It is not "smart", it's just a one-off event that generates the
  components once. No dynamically updating the `Ancestor` component
  to account for runtime changes to the hierarchy.
- It requires an exclusive system. An alternative design could be
  providing a generic system and letting the user add the system with
  their prefered query filter.
- Users need to add the `mark_descendants` system themselves.